minimizer consolidation (#45)
* minimizer consolidation

* pr response
