+++ src/easyscience/fitting/minimizers/minimizer_bumps.py
@@ -3,7 +3,6 @@
 #  © 2021-2023 Contributors to the EasyScience project <https://github.com/easyScience/EasyScience
 
 import copy
-import inspect
 from typing import Callable
 from typing import List
 from typing import Optional
@@ -54,7 +53,6 @@ class Bumps(MinimizerBase):
         :type fit_function: Callable
         """
         super().__init__(obj=obj, fit_function=fit_function, method=method)
-        self._cached_pars_order = ()
         self._p_0 = {}
 
     def available_methods(self) -> List[str]:
@@ -203,90 +201,6 @@ class Bumps(MinimizerBase):
             return _make_func
 
         return _outer(self)
-
-    def _generate_fit_function(self) -> Callable:
-        """
-        Using the user supplied `fit_function`, wrap it in such a way we can update `Parameter` on
-        iterations.
-
-        :return: a fit function which is compatible with bumps models
-        :rtype: Callable
-        """
-        # Original fit function
-        func = self._original_fit_function
-        # Get a list of `Parameters`
-        self._cached_pars_vals = {}
-        for parameter in self._object.get_fit_parameters():
-            key = parameter.unique_name
-            self._cached_pars[key] = parameter
-            self._cached_pars_vals[key] = (parameter.value, parameter.error)
-
-        # Make a new fit function
-        def _fit_function(x: np.ndarray, **kwargs):
-            """
-            Wrapped fit function which now has a bumps compatible form
-
-            :param x: array of data points to be calculated
-            :type x: np.ndarray
-            :param kwargs: key word arguments
-            :return: points calculated at `x`
-            :rtype: np.ndarray
-            """
-            # Update the `Parameter` values and the callback if needed
-            ## TODO clean when full move to new_variable
-            from easyscience.Objects.new_variable import Parameter
-
-            for name, value in kwargs.items():
-                par_name = name[1:]
-                if par_name in self._cached_pars.keys():
-                    ## TODO clean when full move to new_variable
-                    if isinstance(self._cached_pars[par_name], Parameter):
-                        if self._cached_pars[par_name].value != value:
-                            self._cached_pars[par_name].value = value
-                    else:
-                        if self._cached_pars[par_name].raw_value != value:
-                            self._cached_pars[par_name].value = value
-
-                    # update_fun = self._cached_pars[par_name]._callback.fset
-                    # if update_fun:
-                    #     update_fun(value)
-            # TODO Pre processing here
-            for constraint in self.fit_constraints():
-                constraint()
-            return_data = func(x)
-            # TODO Loading or manipulating data here
-            return return_data
-
-        # Fake the function signature.
-        # This is done as lmfit wants the function to be in the form:
-        # f = (x, a=1, b=2)...
-        # Where we need to be generic. Note that this won't hold for much outside of this scope.
-
-        ## TODO clean when full move to new_variable
-        from easyscience.Objects.new_variable import Parameter
-
-        if isinstance(parameter, Parameter):
-            default_value = parameter.value
-        else:
-            default_value = parameter.raw_value
-
-        self._cached_pars_order = tuple(self._cached_pars.keys())
-        params = [
-            inspect.Parameter('x', inspect.Parameter.POSITIONAL_OR_KEYWORD, annotation=inspect._empty),
-            *[
-                inspect.Parameter(
-                    MINIMIZER_PARAMETER_PREFIX + str(name),
-                    inspect.Parameter.POSITIONAL_OR_KEYWORD,
-                    annotation=inspect._empty,
-                    default=default_value,
-                )
-                for name in self._cached_pars_order
-            ],
-        ]
-        # Sign the function
-        _fit_function.__signature__ = inspect.Signature(params)
-        self._fit_function = _fit_function
-        return _fit_function
 
     def _set_parameter_fit_result(self, fit_result, stack_status: bool):
         """
